Moving custom testing predicates to shared header
